requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.5
Pillow==8.0.0
selenium==3.141.0
urllib3==1.25.10